externalized shell commons + readme tweaks
