+++ moo-reports/readme.md
@@ -1,5 +1,5 @@
 # Moo Reports
-=====================
+---------------------
 
 Aggregated Jacoco (code coverage) reports. Generated on `mvn verify`.
