+++ moo-server-socket/readme.md
@@ -1,4 +1,4 @@
 # Moo - chat server
-=====================
+---------------------
 
 Web sockets server for Moo chat service.
