+++ moo-ui-shell/moo-ui-shell-commons/pom.xml
@@ -1,0 +1,11 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>pl.zimowski</groupId>
+    <artifactId>moo-ui-shell</artifactId>
+    <version>1.0.0-SNAPSHOT</version>
+  </parent>
+  <artifactId>moo-ui-shell-commons</artifactId>
+  <name>Moo UI (shell: commons)</name>
+  <description>Common components for shell UI reusable by reader and writer.</description>
+</project>
