+++ moo-ui-shell/moo-ui-shell-commons/readme.md
@@ -1,0 +1,3 @@
+# Moo Shell Commons
+---------------------
+Components and utilities that are reusable between reader and writer.
