+++ moo-ui-shell/moo-ui-shell-reader/pom.xml
@@ -8,4 +8,10 @@
   <artifactId>moo-ui-shell-reader</artifactId>
   <name>Moo UI (shell: reader)</name>
   <description>Text based UI for displaying chat events (read only). In text based console, reading and writing events is most convenient when managed via separate apps.</description>
+  <dependencies>
+    <dependency>
+      <groupId>pl.zimowski</groupId>
+      <artifactId>moo-ui-shell-commons</artifactId>
+    </dependency>  
+  </dependencies>
 </project>
