+++ moo-ui-shell/moo-ui-shell-reader/readme.md
@@ -1,4 +1,4 @@
 # Moo Shell Reader
-=====================
+---------------------
 Console based reader of Moo events. If ran stand alone, allows to view public 
 chats. If ran in combination with shell writer, allows for full chat experience.
