+++ moo-ui-shell/moo-ui-shell-writer/readme.md
@@ -1,3 +1,3 @@
 # Moo Shell Writer
-=====================
+---------------------
 Console based writer of Moo events. Requires a running instance of Moo reader.
