+++ moo-ui-shell/moo-ui-shell-writer/src/main/java/pl/zimowski/moo/ui/shell/writer/EventHandler.java
@@ -7,9 +7,9 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
 import pl.zimowski.moo.api.ClientHandling;
-import pl.zimowski.moo.api.ClientListener;
 import pl.zimowski.moo.api.ServerAction;
 import pl.zimowski.moo.api.ServerEvent;
+import pl.zimowski.moo.ui.shell.commons.AbstractClientListener;
 
 /**
  * Listens for critical server events which are required for 
@@ -19,9 +19,7 @@ import pl.zimowski.moo.api.ServerEvent;
  * @author Adam Zimowski (<a href="mailto:mrazjava@yandex.com">mrazjava</a>)
  */
 @Component
-public class EventHandler implements ClientListener {
-
-	static final Logger LOG = LoggerFactory.getLogger("CHAT_ECHO");
+public class EventHandler extends AbstractClientListener {
 
 	private static final Logger log = LoggerFactory.getLogger(EventHandler.class);
 	
@@ -29,15 +27,16 @@ public class EventHandler implements ClientListener {
 	 * Name reported as author of messages generated by the UI writer.
 	 */
 	static final String AUTHOR = "writer";
-
-	private String nick;
 	
 	@Inject
 	private ClientHandling clientHandler;
 	
-	private String clientId;
 	
-	
+	@Override
+	public String getAuthor() {
+		return AUTHOR;
+	}
+
 	@Override
 	public void onEvent(ServerEvent event) {
 
@@ -68,17 +67,5 @@ public class EventHandler implements ClientListener {
 	@Override
 	public void onConnectToServerError(String error) {
 		LOG.info("({}) could not establish server connection: {}", AUTHOR, error);
-	}
-
-	String getClientId() {
-		return clientId;
-	}
-	
-	String getNick() {
-		return nick;
-	}
-	
-	void setNick(String nick) {
-		this.nick = nick;
 	}
 }
