+++ pom.xml
@@ -261,6 +261,11 @@
         <version>1.0.0-SNAPSHOT</version>
       </dependency>
       <dependency>
+        <groupId>pl.zimowski</groupId>
+        <artifactId>moo-ui-shell-commons</artifactId>
+        <version>1.0.0-SNAPSHOT</version>
+      </dependency>
+      <dependency>
         <groupId>org.slf4j</groupId>
         <artifactId>slf4j-api</artifactId>
         <version>1.7.25</version>
